# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Laravel OpenApi Bitaqty Documentation",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/vendors": {
            "get": {
                "tags": [
                    "VendorProfile"
                ],
                "summary": "Profile",
                "description": "Vendor Profile",
                "operationId": "VendorProfile",
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": ""
                                        },
                                        "title": {
                                            "type": "string",
                                            "example": ""
                                        },
                                        "description": {
                                            "type": "string",
                                            "example": ""
                                        },
                                        "image": {
                                            "type": "string",
                                            "example": ""
                                        },
                                        "data": {
                                            "type": "object",
                                            "example": []
                                        },
                                        "is_read": {
                                            "type": "string",
                                            "example": ""
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "VendorProfile"
                ],
                "summary": "Vendor Update profile",
                "description": "Vendor Update profile",
                "operationId": "VendorUpdateProfile",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": ""
                                    },
                                    "address": {
                                        "type": "string",
                                        "example": ""
                                    },
                                    "phone": {
                                        "type": "string",
                                        "example": ""
                                    },
                                    "commercial_name": {
                                        "type": "string",
                                        "example": ""
                                    },
                                    "email": {
                                        "type": "string",
                                        "example": ""
                                    },
                                    "country_id": {
                                        "type": "integer",
                                        "example": ""
                                    },
                                    "state_id": {
                                        "type": "integer",
                                        "example": ""
                                    },
                                    "currency_id": {
                                        "type": "integer",
                                        "example": ""
                                    },
                                    "image": {
                                        "type": "file",
                                        "format": "file"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Wrong credentials response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": ""
                                        },
                                        "errors": {
                                            "properties": {
                                                "dynamic-error-keys": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string"
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "VendorProfile"
                ],
                "summary": "Vendor Delete Profile",
                "description": "Vendor Delete Profile",
                "operationId": "VendorDeleteProfile",
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Wrong credentials response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": ""
                                        },
                                        "errors": {
                                            "properties": {
                                                "dynamic-error-keys": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string"
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/vendors/changePassword": {
            "put": {
                "tags": [
                    "VendorProfile"
                ],
                "summary": "Vendor Change Password",
                "description": "Change Password Vendor",
                "operationId": "VendorPasswordChange",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "old_password",
                                    "new_password",
                                    "new_password_confirmed"
                                ],
                                "properties": {
                                    "old_password": {
                                        "type": "string",
                                        "example": ""
                                    },
                                    "new_password": {
                                        "type": "string",
                                        "example": ""
                                    },
                                    "password_confirmed": {
                                        "type": "string",
                                        "example": ""
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Wrong credentials response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": ""
                                        },
                                        "errors": {
                                            "properties": {
                                                "dynamic-error-keys": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string"
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "bearerAuth",
                "in": "header",
                "scheme": "bearer"
            }
        }
    },
...
}